+++ test/verify_dry_run.test.sh
@@ -45,7 +45,9 @@ mkdir -p "$THEMES_DIR" "$(dirname "$INDEX_TEMPLATE")" "$(dirname "$PVE_MANAGER_J
 printf '%s\n' 'Proxmox.Utils = { theme_map: {' > "$PROXMOXLIB_JS"
 printf '%s\n' '<script src="/pve2/js/pvemanagerlib.js"></script>' '</head>' '</body>' > "$INDEX_TEMPLATE"
 printf '%s\n' "Ext.define('PVE.form.ViewSelector');" "Ext.define('PVE.tree.ResourceTree');" \
-    "Ext.define('PVE.node.StatusView');" > "$PVE_MANAGER_JS"
+    "Ext.define('PVE.node.StatusView');" "Ext.define('PVE.panel.Config');" \
+    "Ext.define('PVE.sdn.VnetEdit');" "Ext.define('PVE.sdn.SubnetView');" \
+    "Ext.define('PVE.sdn.VnetACLView');" > "$PVE_MANAGER_JS"
 printf '%s\n' '        my $dinfo = df('\''/'\'', 1);' > "$NODES_PM"
 printf '%s\n' 'my $observed = {' '};' > "$PVE_CLUSTER_PM"
 printf '%s\n' 'package PVE::API2;' 'use base qw(PVE::RESTHandler);' '1;' > "$PVE_API2_PM"